fix: refactor hook: useUniqId
