+++ src/components/controls/TextArea/TextArea.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 
+import {useUniqId} from '../../../hooks';
 import {blockNew} from '../../utils/cn';
 import {useForkRef} from '../../utils/useForkRef';
-import {useUniqId} from '../../utils/useUniqId';
 import {ClearButton, mapTextInputSizeToButtonSize} from '../common';
 import {OuterAdditionalContent} from '../common/OuterAdditionalContent/OuterAdditionalContent';
 import type {
