+++ src/hooks/index.ts
@@ -2,3 +2,4 @@ export * from './useBodyScrollLock';
 export * from './useOutsideClick';
 export * from './usePortalContainer';
 export * from './useVirtualElementRef';
+export * from './useUniqId';
