+++ src/hooks/useUniqId/README.md
@@ -1,0 +1,15 @@
+<!--GITHUB_BLOCK-->
+
+# useUniqId
+
+<!--/GITHUB_BLOCK-->
+
+```tsx
+import {useUniqId} from '@gravity-ui/uikit';
+```
+
+The `useUniqId` hook create uniq Id.
+
+## Result
+
+Id. `string`
