+++ src/hooks/useUniqId/index.ts
@@ -1,0 +1,2 @@
+export {useUniqId} from './useUniqId';
+export type {UseUniqIdResult} from './useUniqId';
