+++ src/hooks/useUniqId/useUniqId.ts
@@ -1,7 +1,9 @@
 import React from 'react';
 
-import {NAMESPACE_NEW} from './cn';
-import {getUniqId} from './common';
+import {NAMESPACE_NEW} from '../../components/utils/cn';
+import {getUniqId} from '../../components/utils/common';
+
+export type UseUniqIdResult = string;
 
 function useUniqIdFallback() {
     const idRef = React.useRef<string>();
@@ -15,5 +17,5 @@ function useIdNative() {
     return `${NAMESPACE_NEW}${React.useId()}`;
 }
 
-export const useUniqId: () => string =
+export const useUniqId: () => UseUniqIdResult =
     typeof React.useId === 'function' ? useIdNative : useUniqIdFallback;
